# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NemoReconstruct API",
    "description": "Minimal MOV-to-fVDB reconstruction workflow for Isaac Sim and agent orchestration.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reconstructions/upload": {
      "post": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Upload Reconstruction",
        "operationId": "upload_reconstruction_api_v1_reconstructions_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_reconstruction_api_v1_reconstructions_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reconstructions/{reconstruction_id}": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Reconstruction",
        "operationId": "get_reconstruction_api_v1_reconstructions__reconstruction_id__get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Delete Reconstruction",
        "operationId": "delete_reconstruction_api_v1_reconstructions__reconstruction_id__delete",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/status": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Reconstruction Status",
        "operationId": "get_reconstruction_status_api_v1_reconstructions__reconstruction_id__status_get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/artifacts": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Reconstruction Artifacts",
        "operationId": "get_reconstruction_artifacts_api_v1_reconstructions__reconstruction_id__artifacts_get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionArtifacts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/download/{artifact}": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Download Artifact",
        "operationId": "download_artifact_api_v1_reconstructions__reconstruction_id__download__artifact__get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/retry": {
      "post": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Retry Reconstruction",
        "operationId": "retry_reconstruction_api_v1_reconstructions__reconstruction_id__retry_post",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_reconstruction_api_v1_reconstructions_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "frame_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Rate"
          },
          "sequential_matcher_overlap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequential Matcher Overlap"
          },
          "fvdb_max_epochs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fvdb Max Epochs"
          },
          "fvdb_sh_degree": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fvdb Sh Degree"
          },
          "fvdb_image_downsample_factor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fvdb Image Downsample Factor"
          },
          "splat_only_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Splat Only Mode"
          }
        },
        "type": "object",
        "required": [
          "file",
          "name"
        ],
        "title": "Body_upload_reconstruction_api_v1_reconstructions_upload_post"
      },
      "DeleteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DeleteResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReconstructionArtifacts": {
        "properties": {
          "source_video_url": {
            "type": "string",
            "title": "Source Video Url"
          },
          "splat_ply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Splat Ply Url"
          },
          "scene_usdz_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Usdz Url"
          },
          "sim_bundle_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sim Bundle Url"
          },
          "run_log_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Log Url"
          },
          "metadata_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Url"
          }
        },
        "type": "object",
        "required": [
          "source_video_url"
        ],
        "title": "ReconstructionArtifacts"
      },
      "ReconstructionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pipeline_slug": {
            "type": "string",
            "title": "Pipeline Slug"
          },
          "processing_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Step"
          },
          "processing_pct": {
            "type": "integer",
            "title": "Processing Pct",
            "default": 0
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "source_video_filename": {
            "type": "string",
            "title": "Source Video Filename"
          },
          "frame_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "processing_params": {
            "$ref": "#/components/schemas/ReconstructionParams"
          },
          "artifact_ply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Ply Url"
          },
          "artifact_usdz_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Usdz Url"
          },
          "artifact_bundle_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Bundle Url"
          },
          "artifact_log_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Log Url"
          },
          "artifact_metadata_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Metadata Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "pipeline_slug",
          "source_video_filename",
          "created_at",
          "updated_at"
        ],
        "title": "ReconstructionDetail"
      },
...
      "ReconstructionStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "processing_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Step"
          },
          "processing_pct": {
            "type": "integer",
            "title": "Processing Pct",
            "default": 0
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "updated_at"
        ],
        "title": "ReconstructionStatusResponse"
      },
      "RetryRequest": {
        "properties": {
          "params": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReconstructionParams"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RetryRequest"
      },
...
      "UploadResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pipeline_slug": {
            "type": "string",
            "title": "Pipeline Slug"
          },
          "processing_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Step"
          },
          "processing_pct": {
            "type": "integer",
            "title": "Processing Pct",
            "default": 0
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "source_video_filename": {
            "type": "string",
            "title": "Source Video Filename"
          },
          "frame_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "processing_params": {
            "$ref": "#/components/schemas/ReconstructionParams"
          },
          "artifact_ply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Ply Url"
          },
          "artifact_usdz_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Usdz Url"
          },
          "artifact_bundle_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Bundle Url"
          },
          "artifact_log_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Log Url"
          },
          "artifact_metadata_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Metadata Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "pipeline_slug",
          "source_video_filename",
          "created_at",
          "updated_at"
        ],
        "title": "UploadResponse"
      },
...
    }
  }
}